Syringe Guns: 5% -> 3% über on hit, stock has no reload (#37)
Needleguns are far more chosen than the crossbow because they're great self-defense tools, the Overdose, specifically, is the most chosen out of the three, because its speed bonus is far better utility than, well, nothing. This change makes it so they're still better at self defense than the crossbow, but just not AS good so the crossbow has a chance. The stock needlegun also receives a no reload bonus (attribute 307 is just there to display text when inspecting it, it does nothing) to compete with the Overdose, a change to the Blutsauger to make it more of an offensive weapon is planned, but not very well thought about just yet, maybe later.