Enforcer: crits while disguised, no minicrits while undisguised, crit damage is affected by range (#36)
One of the enforcer's positive stats is that it deals +20% damage while disguised, but that's useless when it's minicrit-boosted while undisguised. This change emphasizes the "shoot while disguised" bit, since it's far easier to hit once while disguised than hitting a headshot with the ambassador for example, the crit falloff attribute that was oringinally used by the ambassador was added to this weapon instead.